Update `Callback` type synonym for node callbacks
Given that values of this type are only ever used in FFI code, there's
no need for a type synonym; the right hand side of the previous type
synonym didn't have any effect on whether the code type checked.

This is a little surprising, so I thought a foreign data declaration
would be more appropriate. I also added a comment, because I was
previously confused by what this type actually represents.