Scan for mtimes recursively.
Testing the mtime of a directory to find the latest changed file inside
it doesn't work on OSX's file system, and likely elsewhere.

This can be slow, and is duplicated by `psc`. If we could figure out
whether `psc` actually built something, that would be preferable, but
currently the only obvious solution would be to scan the `output`
folder, which is equally slow.

Closes #105.