Fix searching in parent directories on Windows
When looking for a file in this or parent directories, previously, pulp
would give up once we reach "/". This won't work on Windows, as the root
directory is usually not called "/".

Instead, stop looking at parents as soon as the parent directory of the
current directory is the same as the current directory. So for example,
this should work, regardless of whether the root is "/", "C:/", "D:/",
and so on.