Fix pulp server, fixes #151
`pulp server` now uses the same mechanisms as `pulp build` internally
so that stuff is less likely to break. This also allows us to drop a few
dependencies, which is nice.

The options for `pulp server` have been adjusted appropriately - that
is, --no-info and --config FILE have been removed (so this is a breaking
change).

Also made `pulp --watch server` just act the same as `pulp server` (as
`pulp server` already watches the source directories).